Remove `include/grpcpp/impl/codegen/server_interceptor.h` (#31281)
